+++ packages/lit-dev-content/samples/tutorials/wc-to-lit/06/before/index.html
@@ -2,11 +2,6 @@
 <html>
  <head>
    <script src="./rating-element.js" type="module"></script>
-   <style>
-     span {
-       border: 1px solid red;
-     }
-   </style>
  </head>
  <body>
    <template id="rating-element-template">
