+++ packages/lit-dev-content/samples/tutorials/wc-to-lit/09/before/project.json
@@ -1,7 +1,7 @@
 {
   "extends": "/samples/base.json",
   "files": {
-    "index.html": {},
-    "rating-element.ts": {}
+    "rating-element.ts": {},
+    "index.html": {}
   }
 }
