--- packages/lit-dev-content/samples/tutorials/wc-to-lit/12/after/index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <script src="./rating-element.js" type="module"></script>
-  </head>
-  <body>
-    <rating-element rating="5"></rating-element>
-  </body>
-</html>
