--- packages/lit-dev-content/samples/tutorials/wc-to-lit/13/after/project.json
@@ -1,7 +1,0 @@
-{
-  "extends": "/samples/base.json",
-  "files": {
-    "index.html": {},
-    "rating-element.ts": {}
-  }
-}
