+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/00.md
@@ -16,8 +16,10 @@ Lit provides APIs to simplify common web components tasks like managing properti
 * How to build a Lit component
 
 <aside class="warning">
+
 This tutorial focuses on the web components specifications and how Lit works under the hood to help  you build web components more easily.
-For a more general introduction to Lit, take a look at the <a href="https://lit.dev/docs/">Lit docs</a> or other <a href="https://lit.dev/tutorials/">tutorials</a>.
+For a more general introduction to Lit, take a look at the [Lit docs](/docs/) or [other tutorials](/tutorials/).
+
 </aside>
 
 ## What you’ll build
@@ -26,5 +28,7 @@ For a more general introduction to Lit, take a look at the <a href="https://lit.
 * A thumbs up / down LitElement-based web component
 
 <aside class="positive">
-Already know the basics of web components? Skip to <a href="#09">Lit templates</a>, to see how Lit can help!
+
+Already know the basics of web components? Skip to [Lit templates](#09), to see how Lit can help!
+
 </aside>
