+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/01.md
@@ -11,7 +11,7 @@ You've likely already used the ES modules specification, which allows you to cre
 
 ## Defining a Custom Element
 
-The custom elements specification lets users define their own HTML elements using JavaScript. The names must contain a hyphen (`-`) to differentiate them from native browser elements. Clear the `rating-element.<litdev-code-language-display></litdev-code-language-display>` file and define a custom element class:
+The custom elements specification lets users define their own HTML elements using JavaScript. The names must contain a hyphen (`-`) to differentiate them from native browser elements. Clear the <code>rating-element.<litdev-code-language-display></litdev-code-language-display></code> file and define a custom element class:
 
 ### rating-element.<litdev-code-language-display></litdev-code-language-display>
 
@@ -31,7 +31,7 @@ customElements.define('rating-element', RatingElement);
 
 {% endswitchable-sample %}
 
-A custom element is defined by associating a class that extends `HTMLElement` with a hyphenated tag name. The call to `customElements.define` tells the browser to associate the class `RatingElement` with the tagName `'rating-element'`. This means that every element in your document with the name `<rating-element>` will be associated with this class.
+You define a custom element by associating an element class with a tag name. The class must extend `HTMLElement`, and the tag name must must contain a hyphen (`-`) to differentiate it from a native browser elements.
 
 <aside class="warning">
 Since <code>customElements.define</code> has a global scope, you cannot currently call it more than once for the same tag name even if the same class definition is given.
@@ -47,7 +47,9 @@ Place a `<rating-element>` in the document body and see what renders.
 </body>
 ```
 
-Now, looking at the output, you'll see that nothing has rendered. This is expected, because you haven't told the browser how to render `<rating-element>`. You can confirm that the custom element definition has succeeded by right-clicking on the *result preview window*, selecting "inspect" or "inspect element", and checking the constructor of the `<rating-element>` by calling the following expression in the console:
+Looking at the output, you'll see that nothing has rendered. This is file; you haven't told the browser how to render the element yet.
+
+You can confirm that the custom element definition has succeeded by right-clicking on the *result preview window*, selecting "inspect" or "inspect element", and checking the constructor of the `<rating-element>` by calling the following expression in the console:
 
 ```text
 document.querySelector('rating-element').constructor
