+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/02.md
@@ -8,9 +8,9 @@ Custom elements come with a set of lifecycle hooks. In this section you'll use t
 The `constructor` is called when the element is first created: for example, by calling `document.createElement(‘rating-element')` or `new RatingElement()`. The constructor is a good place to set up your element.
 
 <aside class="negative">
-<p>It is considered bad practice to do DOM manipulations in the constructor, because DOM manipulations can slow down initial load times.</p>
 
-<p>Additionally, the browser will throw an error if a custom element definition that manipulates the DOM in the <code>constructor</code> is loaded while the browser is still parsing the document because the parser does not know how to handle changes to the DOM mid-parse. This is common in inline scripts or in a <code>&lt;script></code> that is not <code>type="module"</code></p>
+It's bad practice to do DOM manipulations in the `constructor`, because DOM manipulations can slow down initial load times, and [cause some problems in some edge cases](/playground/#gist=34f89b3c8c1c930b3d943c49656d1f2a).
+
 </aside>
 
 The `connectedCallback` is called when the custom element is attached to the DOM. This is typically where initial DOM manipulations happen.
