+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/03.md
@@ -1,6 +1,6 @@
 ## Why Shadow DOM?
 
-In the previous step, you'll notice that the selectors in the `<style>` tag that you inserted select any `<rating-element>` on the page as well as any `<button>`. This may result in the styles leaking out of the element and selecting other nodes that you may not intend to style. Additionally, other styles outside of this custom element may unintentionally style the nodes inside your custom element.
+You may have noticed that the styles you added in the previous step select any `<rating-element>` on the page as well as any `<button>`. This can result in the styles leaking out of the element and selecting other nodes that you may not intend to style. And styles **outside** of your custom element may unintentionally style the nodes **inside** the element.
 
 For example, try putting a style tag in the head of the main document:
 
@@ -84,5 +84,7 @@ An experiment: add a node as a direct child of the `<rating-element>`:
 Once the page refreshes, you'll see that the new `<div>` does not show up. This is because shadow DOM has features to control how child nodes—sometimes called *light DOM*—are rendered or *projected* into the shadow DOM.
 
 <aside class="info">
-This tutorial doesn't cover light DOM projection, but you can learn more about it in this <a href="https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_templates_and_slots#adding_flexibility_with_slots">MDN article</a> or the <a href="https://lit.dev/docs/components/shadow-dom/#slots">Lit docs site</a>.
+
+This tutorial doesn't cover light DOM projection, but you can learn more about it in [this article](https://web.dev/shadowdom-v1/#composition-and-slots) or the [Lit documentation](/docs/components/shadow-dom/#slots).
+
 </aside>
