+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/04.md
@@ -1,6 +1,6 @@
 ## Why Templates
 
-Using `innerHTML` and template literal strings with no sanitization may cause security issues with script injection. In the past, to create HTML templates it was common to use `DocumentFragments` or scripts with specialized `type` attibutes, but these solutions had issues such as images loading or scripts running when the templates are defined but not rendered, and obstacles for reusability across development environments.
+Using `innerHTML` and template literal strings with no sanitization may cause security issues with script injection. In the past, developers used various workarounds to implement HTML templates, but all of these workarounds had issues.
 
 This is where the `<template>` element comes in; templates provide truly *inert* DOM, a highly performant method to clone nodes, and reusable templating.
 
@@ -74,9 +74,9 @@ connectedCallback() {
 To use this `<template>` element:
 
 1. Query the template.
-2. Get it's contents.
+2. Get its contents.
 3. Clone those nodes with `templateContent.cloneNode`
    * The `true` argument performs a deep clone.
 4. Initialize the dom with the data.
 
-Congratulations, you now have a web component! Unfortunately it does not do anything yet, so next, add some functionality.
+Congratulations, now your web component's DOM is encapsulated, but the DOM is still static. In the next steps, you'll add support for updating the rating.
