+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/05.md
@@ -53,8 +53,9 @@ get rating() {
 {% endswitchable-sample %}
 
 1. Remove the `constructor`.
-2. Add a setter and getter for the `rating` property.
-3. Update the `rating` element's text if it's available.
+2. Declare a private `_rating` property set to `0`.
+3. Add a setter and getter for the `rating` property.
+4. Update the `rating` element's text if it's available.
 
 This means if you were to set the `rating` property on the element, the view will update; give it a quick test in your Dev Tools console! Select the `<rating-element>` and set the `rating` property to `5`.
