+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/07.md
@@ -1,20 +1,29 @@
 ## Styling Buttons With Attributes
 
-Now all that's missing is the button functionality. The behavior of this component should allow the user to provide a single up or down vote rating and give visual feedback to the user. You can implement this with some event listeners and a reflecting the new value to the attribute, but first update the styles to give visual feedback by appending the following lines:
+Now all that's missing is the button functionality. This component should allow the user to provide a single up or down vote rating and give visual feedback to the user. In this step you'll do some setup work:
+
+* Add a new `vote` property and attribute, like the `rating` property you added previously.
+* Add new styles for the vote-up and vote-down buttons based on the current value of the `vote` attribute.
+
+Then in the following step you'll add the event listeners to handle button clicks and set the `vote` attribute.
+
+Start by adding the following lines:
 
 ### index.html
 
 ```css
-<style>
-  ...
+<template>
+  <style>
+    ...
 
-  :host([vote=up]) .thumb_up {
-    fill: green;
+    :host([vote=up]) .thumb_up {
+      fill: green;
+    }
+    :host([vote=down]) .thumb_down {
+    fill: red;
   }
-  :host([vote=down]) .thumb_down {
-   fill: red;
- }
-</style>
+  </style>
+</template>
 ```
 
 In shadow DOM the `:host` selector refers to the node or custom element that the shadow root is attached to.
@@ -105,7 +114,9 @@ export class RatingElement extends HTMLElement {
 You initialize the `_vote` instance property with `null` as a class-member property, and in the setter you check if the new value is different. If so, you adjust the `rating` accordingly and, importantly, reflect the `vote` attribute back to the host with `this.setAttribute`.
 
 <aside class="warning">
-Note, this is not the most-efficient way of writing this, but it is meant to exemplify how state works on a web component.
+
+It is not recommended to manipulate `rating` inside the `vote` setter in this way as it's not the most efficient way to update `rating`, but it's the most convenient way for this tutorial.
+
 </aside>
 
 ## Handle "vote" Attribute Changes
