+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/08.md
@@ -86,11 +86,11 @@ export class RatingElement extends HTMLElement {
 
 In this step you:
 
-* Bind some click listeners to the element (`.bind(this)`) and keep the references around as private class members
-* Listen for `click` events on the `<button>`s in `connectedCallback`
-* Clean up these listeners in `disconnectedCallback`
-* Set `this.vote` appropriately in the click listeners themselves
+* Bind some click listeners to the element (`.bind(this)`) and keep the references around as private class members.
+* Listen for `click` events on the `<button>`s in `connectedCallback`.
+* Clean up these listeners in `disconnectedCallback`.
+* Set `this.vote` appropriately in the click listeners themselves.
 
 Congratulations, you now have a fully-featured web component; try clicking on some buttons!
 
-The issue now is that my JS file is now reaching 103 lines (99 in TS), my HTML file 38 lines, and the code is quite verbose, imperative, and unreadable for such a simple component. This is where Lit comes in!
+The bad news now is that the element definition is around 100 lines, the HTML file is 38 lines, and the code is quite verbose, imperative, and unreadable for such a simple component. This is where Lit comes in!
