+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/09.md
@@ -8,12 +8,12 @@
 
   In summary, we have built a web component that:
 
-  * Accepts a `vote="up"` or `vote="down"` attribute
-  * Accepts a `rating="0"` attribute
-  * Accepts those attribtues as properties
-  * Reflects the `vote` property to the `vote` attribute
-  * Styles the buttons based on the `vote` attribute
-  * Sets the `rating` and `vote` to the appropriate values on click
+  * Accepts a `vote="up"` or `vote="down"` attribute.
+  * Accepts a `rating="0"` attribute.
+  * Accepts those attributes as properties.
+  * Reflects the `vote` property to the `vote` attribute.
+  * Styles the buttons based on the `vote` attribute.
+  * Updates the `rating` and `vote` values when the user clicks one of the buttons.
 </details>
 
 ---
@@ -27,7 +27,9 @@ Template elements also lend themselves to imperative code. In many cases this le
 This is where Lit templates comes in! Lit lets you write templates in JavaScript, then efficiently render and re-render those templates together with data to create and update DOM. It is similar to JSX and Virtual DOM libraries but it runs natively in the browser and much more efficiently than VDOM in many cases.
 
 <aside class="info">
-You can learn about more Lit's templating features that we will not cover (such as template repeating and property binding) on the <a href="https://lit.dev/docs/templates/overview/">Lit docs' templating section</a>.
+
+This tutorial uses only a few of the features supported by Lit templates.  For full coverage, see [Templates](/docs/templates/overview/) in the Lit documentation.
+
 </aside>
 
 ## Using Lit templates
@@ -38,12 +40,10 @@ Lit then uses `<template>` elements under the hood to provide fast rendering as 
 
 Start by migrating the `<template>` in index.html into a Lit template by adding a `render()` method to the web component:
 
-### index,js
-
-{% switchable-sample %}
+### index.<litdev-code-language-display></litdev-code-language-display>
 
 ```ts
-// Dont forget to import from Lit!
+// Don't forget to import from Lit!
 import {render, html} from 'lit';
 
 export class RatingElement extends HTMLElement {
@@ -85,62 +85,14 @@ export class RatingElement extends HTMLElement {
   }
 }
 ```
-
-```js
-// Dont forget to import from Lit!
-import {render, html} from 'lit';
-
-export class RatingElement extends HTMLElement {
-  ...
-  render() {
-    if (!this.shadowRoot) {
-      return;
-    }
-
-    const template = html`
-      <style>
-        :host {
-          display: inline-flex;
-          align-items: center;
-        }
-        button {
-          background: transparent;
-          border: none;
-          cursor: pointer;
-        }
-
-       :host([vote=up]) .thumb_up {
-         fill: green;
-       }
-
-       :host([vote=down]) .thumb_down {
-         fill: red;
-       }
-      </style>
-      <button class="thumb_down">
-        <svg xmlns="http://www.w3.org/2000/svg" height="24" viewbox="0 0 24 24" width="24"><path d="M15 3H6c-.83 0-1.54.5-1.84 1.22l-3.02 7.05c-.09.23-.14.47-.14.73v2c0 1.1.9 2 2 2h6.31l-.95 4.57-.03.32c0 .41.17.79.44 1.06L9.83 23l6.59-6.59c.36-.36.58-.86.58-1.41V5c0-1.1-.9-2-2-2zm4 0v12h4V3h-4z"/></svg>
-      </button>
-      <span class="rating">${this.rating}</span>
-      <button class="thumb_up">
-        <svg xmlns="http://www.w3.org/2000/svg" height="24" viewbox="0 0 24 24" width="24"><path d="M1 21h4V9H1v12zm22-11c0-1.1-.9-2-2-2h-6.31l.95-4.57.03-.32c0-.41-.17-.79-.44-1.06L14.17 1 7.59 7.59C7.22 7.95 7 8.45 7 9v10c0 1.1.9 2 2 2h9c.83 0 1.54-.5 1.84-1.22l3.02-7.05c.09-.23.14-.47.14-.73v-2z"/></svg>
-      </button>`;
-
-    render(template, this.shadowRoot);
-  }
-}
-```
-
-{% endswitchable-sample %}
 
 You may also delete your `<template>` from `index.html`.
 
-In this `render` method you define a variable called `template` and invoke the `html` tagged template literal function. You will also notice that you have performed a simple data binding inside the `span` element by using something called *template literal interpolation* (`${...}`).
+In this `render()` method you define a variable called `template` and invoke the `html` tagged template literal function. Also notice the text `${this.rating}` inside the `span` element. This is a Lit _expression_, and it takes the place of setting the span's `innerText` imperatively.
 
-This means that you will eventually no longer need to imperatively update that node and can remove the `.innerText` assignment in your `connectedCallback`.
+Additionally, you call the Lit `render()` method which synchronously renders the template into the shadow root. Every time the component's `render()` method is called, the rating value will be updated.
 
-Additionally, you call the lit render method which synchronously renders the template into the shadow root.
-
-Now, call `this.render()` in the `connectedCallback` and remove the logic related to inserting the template:
+Eventually you won't have to call `render()` imperatively at all. For now, call `this.render()` in the `connectedCallback` and remove the logic related to inserting the template and setting the `innerText` on the `.rating` span:
 
 
 ### rating-element.<litdev-code-language-display></litdev-code-language-display>
@@ -176,3 +128,5 @@ connectedCallback() {
 ```
 
 {% endswitchable-sample %}
+
+Now all your logic and templating is packaged in a single location! In the next step, you will clean up all the imperative code by moving it to the template.
