+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/10.md
@@ -1,6 +1,6 @@
 ## Event Listeners
 
-Now that you have gotten rid of the `<template>` element in `index.html`, refactor the code to take advantage of `lit-html`s features in the newly-defined `render` method. You can start by leveraging Lit's event listener binding syntax:
+Now that you have gotten rid of the `<template>` element in `index.html`, refactor the code to take advantage of Lit template features in the newly-defined `render()` method. You can start by leveraging Lit's event listener binding syntax:
 
 ### rating-element.<litdev-code-language-display></litdev-code-language-display>
 
@@ -17,15 +17,17 @@ Now that you have gotten rid of the `<template>` element in `index.html`, refact
 Lit templates can add an event listener to a node with the `@EVENT_NAME` binding syntax where, in this case, you update the `vote` property every time these buttons are clicked.
 
 <aside class="info">
-You can learn more about the Lit binding syntax on the <a href="https://lit.dev/docs/templates/expressions/">Lit docs' Expressions section</a>.
+
+You can learn more about the Lit binding syntax in the [Lit Expressions documentation](/docs/templates/expressions/).
+
 </aside>
 
 Next:
 
-* Clean up the `_boundOn[Up|Down]Click` class members
-* Clean up the event logic in `connectedCallback`
-* Remove `disconnectedCallback`
-* Remove the `_on[Up|Down]Click` methods
+* Clean up the `_boundOn[Up|Down]Click` class members.
+* Clean up the event logic in `connectedCallback`.
+* Remove `disconnectedCallback`.
+* Remove the `_on[Up|Down]Click` methods.
 
 ### rating-element.<litdev-code-language-display></litdev-code-language-display>
 
@@ -63,12 +65,12 @@ export class RatingElement extends HTMLElement {
 
 {% endswitchable-sample %}
 
-What did we achieve? You were able to remove:
+You were able to remove:
 
-* The click listener logic from all three callbacks
-* The `disconnectedCallback` entirely!
-* All of the DOM initialization code from the `connectedCallback` making it look much more elegant
-* `_onUpClick` and `_onDownClick` listener methods
+* The click listener logic from all three callbacks.
+* The `disconnectedCallback` entirely.
+* All of the DOM initialization code from the `connectedCallback` making it look much more elegant.
+* `_onUpClick` and `_onDownClick` listener methods.
 
 ## Rendering Property Updates
 
@@ -106,6 +108,8 @@ set rating(value) {
   this.render();
 }
 
+...
+
 set vote(newValue) {
   ...
 
@@ -119,13 +123,15 @@ set vote(newValue) {
 {% endswitchable-sample %}
 
 <aside class="warning">
-Note, this is not the most-efficient way of writing this, but its complexity is meant to exemplify features of the web component and Lit lifecycle.
+
+Note, synchronously calling `render()` in the setters of `rating` and `vote` is not the most efficient way to update the DOM, but it is a good way to illustrate where `LitElement` calls `render()` (next step).
+
 </aside>
 
 Here, you:
 
-* Removed the dom update logic from the `rating` setter
-* Added a call to `render` from the `vote` setter
+* Removed the dom update logic from the `rating` setter.
+* Added a call to `render` from the `vote` setter.
 
 Now the template is much more readable as you now can see where the bindings and event listeners are applied.
