+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/11.md
@@ -1,10 +1,9 @@
 ## Why LitElement
 
-We still have some problems with our code:
+There are still ways the code can be improved:
 
-1. If you change the `vote` property or attribute, it may change the `rating` property which will result in calling `render` twice.
-   * Despite repeat calls of `render` being a *very* efficient, the browser is still spending time calling that function twice
-2. It is tedious adding new properties and attributes as it requires a lot of boilerplate code
+1. If the `vote` property or attribute is changed, the `rating` property may also be changed which will result in calling `render` twice. Despite repeat calls of `render` being a *very* efficient, the browser is still needlessly calling that function twice.
+2. It is tedious adding new properties and attributes as it requires a lot of boilerplate code.
 
 This is where `LitElement` comes in!
 
@@ -12,7 +11,12 @@ This is where `LitElement` comes in!
 
 ## Defining A LitElement
 
-Start by importing and subclassing the `LitElement` base class from the `lit` package and in TS, use the `@customElement` decorator to replace the `customElements.define` call:
+Start by importing and subclassing the `LitElement` base class from the `lit` package<litdev-code-language-display>
+  <span slot="js">.</span>
+  <span slot="ts">,
+    and use the `@customElement` decorator to replace the `customElements.define` call.
+  </span>
+</litdev-code-language-display>
 
 ### rating-element.<litdev-code-language-display></litdev-code-language-display>
 
@@ -45,12 +49,19 @@ export class RatingElement extends LitElement {
 
 {% endswitchable-sample %}
 
-You import `LitElement` which is the new base class for the `<rating-element>`.
+What's changed?
 
-Next you keep the `html` import and add `css` to your imports which allows us to define css tagged template literals which enable math, templating, and other features under the hood.
+* `LitElement` is imported.
+* `LitElement` is now the new base class for `<rating-element>`.
+* `html` is still imported.
+* `css` is imported.
+
+The `css` tag function allows you to define CSS tagged template literals, which enable math, templating, and other features under the hood.
 
 <aside class="info">
-For a comprehensive list of styling features, please visit the <a href="https://lit.dev/docs/components/styles/">Lit docs's styling section</a>.
+
+For a comprehensive list of styling features, Refer to the [Lit Styles documentation](https://lit.dev/docs/components/styles/).
+
 </aside>
 
 ## Styling With Lit
@@ -117,11 +128,11 @@ export class RatingElement extends LitElement {
 
 This is where most styles live in Lit.
 
-Lit will take these styles and use browser features such as [*Constructable Stylesheets*](https://developers.google.com/web/updates/2019/02/constructable-stylesheets) to provide faster rendering times as well as pass it through the web components polyfill on older browsers if necessary.
+Where available, Lit uses browser features such as [*Constructable Stylesheets*](https://web.dev/constructable-stylesheets/) to provide faster rendering times for your styles.
 
 ## Render
 
-Next, `render` is a `LitElement` lifecycle method which requires us to **return** a Lit template and no longer needs to check for the shadow root:
+Next, `render()` is a `LitElement` lifecycle method which requires us to **return** a Lit template and no longer needs to check for the shadow root:
 
 ### rating-element.<litdev-code-language-display></litdev-code-language-display>
 
@@ -167,4 +178,4 @@ render() {
 
 {% endswitchable-sample %}
 
-Right now, you'll have a stylable, element that doesn't react to property and attribute changes. Let's fix that!
+Now you have a styled element! In the next step you will make the element react to property and attribute changes.
