+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/12.md
@@ -9,7 +9,9 @@
 Lit introduces a set of render lifecycle callback methods on top of the native web component callbacks. These callbacks are triggered when declared Lit reactive properties are changed.
 
 <aside class="info">
-To learn more about the Lit rendering lifecycle, please visit the <a href="https://lit.dev/docs/components/lifecycle/">Lit docs' lifecycle section</a>.
+
+To learn more about the Lit reactive update cycle, see the [Lit Lifecycle documentation](/docs/components/lifecycle/).
+
 </aside>
 
 To use this feature, you must statically declare which properties are *Reactive Properties* – properties that will trigger the render lifecycle when changed:
@@ -59,49 +61,24 @@ export class RatingElement extends LitElement {
 
 Here, you:
 
-* Declare that `rating` and `vote` are reactive properties
-  * They will trigger the `LitElement` rendering lifecycle when changed
-* Define the types that will be used to convert the `string` attributes into properties
-* Remove setter and getter logic for the reactive properties
-* Remove the attribute handling logic
-* Clean up some uneccessary private class members
+* Declare that `rating` and `vote` are reactive properties.
+  * They will trigger the `LitElement` rendering lifecycle when changed.
+* Define the [type]/docs/components/properties/#property-options) that will be used to convert the `string` attributes into properties.
+* Remove setter and getter logic for the reactive properties.
+* Remove the attribute handling logic.
+* Clean up some unnecessary private class members.
 
-<aside class="info">
-  If you want to pass a complex object into a <code>LitElement</code>, you may define it as <code>type: Object</code>, but often good practice is not pass objects as a string attribute but rather as a property and set <code>attribute: false</code>.
+<aside class="positive">
 
-  <details>
-    <summary>Example:</summary>
+It is generally good practice to pass complex objects as [properties](/docs/templates/expressions/#property-expressions) rather than [attributes](/docs/templates/expressions/#attribute-expressions). Read more on reactive property attribute conversion in the [Lit documentation](/docs/components/properties/#attributes).
 
-```js
-document.querySelector(‘user-profile').user = this.user;
-```
-  </details>
-  <details>
-    <summary><code>lit-html</code> Example:</summary>
-
-```html
-<user-profile .user=${this.user}></user-profile>
-```
-
-  Even better practice is to spread out the object onto the web component declaratively. E.g.
-
-```html
-<user-profile .name=${this.user.name} .age=${this.user.age}>
-  ${this.user.family.map(member => html`
-        <family-member
-             .name=${member.name}
-             .relation=${member.relation}>
-        </family-member>`)}
-</user-profile>
-```
-</details>
 </aside>
 
 Additionally, the `reflect` flag on the `vote` property will automatically update the host element's `vote` *attribute* that you formerly updated in the `vote` setter.
 
 ## willUpdate Lifecycle Callback
 
-Now update the `rating` when the `vote` property changes in the `willUpdate` Lit lifecycle method:
+Now update the `rating` when the `vote` property changes in the `willUpdate()` Lit lifecycle method:
 
 ### rating-element.<litdev-code-language-display></litdev-code-language-display>
 
@@ -158,8 +135,8 @@ willUpdate(changedProps) {
 
 {% endswitchable-sample %}
 
-The logic here is the same from the `vote` setter logic just moved to to the `willUpdate` lifecycle method.
+The logic here is the same from the `vote` setter logic just moved to to the `willUpdate()` lifecycle method.
 
-The `willUpdate` method is called before `render` every time reactive property is changed, because LitElement batches property changes and makes rendering asynchronous. Changes to reactive properties (like `this.rating`) in `willUpdate` will not trigger unnecessary render lifecycle calls.
+The `willUpdate()` method is called before `render()` every time reactive property is changed, because LitElement batches property changes and makes rendering asynchronous. Changes to reactive properties (like `this.rating`) in `willUpdate()` will not trigger unnecessary render lifecycle calls.
 
 Congratulations, you should have a working Lit Element!
