+++ packages/lit-dev-content/site/tutorials/content/wc-to-lit/13.md
@@ -2,26 +2,32 @@
 
 You've built a "vanilla" web component and converted it to a Lit-based web component!
 
-We've been able to convert an unreadable 93-line vanilla web component with a 43-line html template to a 68-line Lit-based web component template-and-all! (JS: 104 + 43 lines to 75 lines)
+You've also converted an unreadable <litdev-code-language-display>
+  <span slot="js">101</span>
+  <span slot="ts">99</span>
+</litdev-code-language-display>-line vanilla web component with a 27-line html template to a <litdev-code-language-display>
+  <span slot="js">75</span>
+  <span slot="ts">68</span>
+</litdev-code-language-display>-line Lit-based web component template-and-all!
 
-## What's Does lit-html Do again?
+## What's do Lit templates do again?
 
-`lit-html`, Lit's rendering system does the following:
+Lit's rendering system does the following:
 
-* Efficiently renders and updates DOM
-* Enables you to easily package your template with your component
-* Allows you to write templates without the need for a compiler
-* Enables you to write readable, declarative templates and code with special binding syntax
+* Efficiently renders and updates DOM.
+* Enables you to easily package your template with your component.
+* Allows you to write templates without the need for a compiler.
+* Enables you to write readable, declarative templates and code with special binding syntax.
 * And much more!
 
 ## What does LitElement do again?
 
 `LitElement`, Lit's web component base class does the following:
 
-* Handles converting attributes to properties
-* Handles reflecting properties back into attributes
-* Provides a render and re-render reactive lifecycle
-* Handles your re-renders efficiently and asynchronously
+* Handles converting attributes to properties.
+* Handles reflecting properties back into attributes.
+* Provides a render and re-render reactive lifecycle.
+* Handles your re-renders efficiently and asynchronously.
 * And does all of this without a build and only ~5kb lines of code!
 
 `LitElement` also enables many more advanced features as the Lit team & community has been building an ecosystem of tooling around it. Learn more by reading the [Lit Documentation](https://lit.dev/docs/)!
